[Feat] View 전체 수정된 컨벤션 반영
